requirements.txt
httpx
python-dotenv
uv
